fix: correct TagIds and TrackerEntry.Downloaded data model types
TagIds: migration 020 stored tag IDs as JSON array but the C# model kept
int, so Dapper would read the JSON string as 0. Change to List<int> and
register EmbeddedDocumentConverter<List<int>> in DbFactory.

TrackerEntry.Downloaded: int overflows at ~2GB. Change to long in both
the domain model and API resource. SQLite already stores integers with
sufficient width; migration 021 is a documented no-op for SQLite.

+++ src/NzbDrone.Core/Torrents/TrackerEntry.cs
@@ -12,7 +12,7 @@ public class TrackerEntry : ModelBase
     public bool Enabled { get; set; }
     public int Seeders { get; set; }
     public int Leechers { get; set; }
-    public int Downloaded { get; set; }
+    public long Downloaded { get; set; }
     public int TotalAnnounces { get; set; }
     public int SuccessfulAnnounces { get; set; }
     public int ConsecutiveFailures { get; set; }
